* basic combat implemented

+++ SuddenAttack/Assets/Scripts/Model/Data/UnitData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class UnitData : ScriptableObject
 {
-    public int HitPoints { get; set; }
+    public float HitPoints { get; set; }
     public float MoveSpeed { get; set; }
     public float FireSpeed { get; set; }
     public float Damage { get; set; }
